create Entity class and continue class diagram
